Refactor gooey menu, based on Chakra UI
